SiStripMonitorCondData plugin: use the helper class with the same functionality (minimal changes to support both)
